api giao viec

+++ controllers/nhanvien_controller.js
@@ -1,3 +1,4 @@
+const { nhanvien_congviecModel, nhanvien_congviec_model } = require("../models/nhanvien_congviec_models");
 const { nhanvienModel } = require("../models/nhanvien_model");
 
 exports.addnhanvien = async (req, res, next) => {
@@ -57,11 +58,20 @@ exports.updatenhanvien = async (req, res, next) => {
 exports.deletenhanvien = async (req, res, next) => {
     try {
         let id = req.params.id;
+        
+        // Kiểm tra xem id nhân viên có tồn tại trong bảng NhanVien_cong viec hay không
+        const existingAssignment = await nhanvien_congviec_model.findOne({ id_nhanvien: id });
+        if (existingAssignment) {
+            return res.json({ status: 403, message: "Nhân viên này đã có công việc được gán, không thể xóa." });
+        }
+
+        // Nếu không tồn tại trong bảng NhanVien_cong viec, tiến hành xóa
         let result = await nhanvienModel.findByIdAndDelete(id);
-        res.json({ status: "delete thanh cong", result: result });
+        res.json({ status: 200, message: "Xóa nhân viên thành công", result: result });
     } catch (error) {
-        res.json({ status: "delete khong thanh cong", result: error });
+        res.json({ status: "failed", message: "Xóa nhân viên không thành công", result: error });
     }
+
 };
 //getnhanvienbyid
 exports.getnhanvien = async (req, res, next) => {
